requirements.txt
flask
pylint 
pytest
gunicorn
pytest-cov
mysql-connector-python
opensearch-py
pytest-cov
psycopg2-binary
certbot
